make scene tab update on tumbnail update
